requirements.txt
django==3.2.4
python-decouple==3.4
djangorestframework==3.12.4
ipython==7.24.1
django-crispy-forms==1.11.2